fix: resolve 34 bugs across security, state management, hooks, and services
Security (critical/high):
- Restrict SQL commands to the sql window via Tauri window-label gating
- Add allowlist for run_gh_command (only pr, auth subcommands)
- Fix AppleScript injection in macOS notifications (escape \ and ")
- Fix PowerShell/cmd injection in Claude launcher (escape special chars)
- Migrate credentials from plaintext settings.json to OS keychain (keyring crate)
- Remove style from DOMPurify ALLOWED_ATTR to prevent CSS-based XSS

State management & components:
- Use content-aware cache key in pr-store (PR numbers + statuses, not just length)
- Deep merge in updateSettings to prevent nested object destruction
- Auto-persist updateSettings with debounced save
- Add key prop to PRDetailPanel to reset state on PR switch
- Fix SettingsFlyout spurious saves by clearing timerRef after flush
- Fix FloatingBadge stale closure with ref pattern
- Guard restorePersistedSection from overwriting user navigation
- Preserve hasCompletedInitialLaunch in initStore.reset

Hooks:
- Fix useAdoPolling excessive re-runs via settingsRef pattern (scalar deps only)
- Fix async Tauri listener leak in 6 hooks (cancelled-flag pattern)
- Add dependency array to useKeyboardNav effect
- Add .catch() to useSyntaxHighlight promise chain

Services:
- Distinguish rate-limit 403 from auth 403 (new GitHubRateLimitError class)
- Recover from 304 with missing cache entry (retry without ETag)
- Add auth error handling to getRaw, post, put, patch methods
- Fix PollingManager stop()+pollNow() race with _stopped flag
- Remove dead since param from getClosedPRs (GitHub ignores it)
- Cap ETag cache at 500 entries with LRU eviction
- Fix ADO executeQuery for tree/oneHop query types
- Fix avgWaitHours denominator (count reviews with timestamps only)
- Invalidate contributor-stats cache on config change
- Use @tauri-apps/api/path join for cross-platform temp file paths

Rust backend:
- Wrap cache_save_prs in transaction (prevent data loss on crash)
- Change log rotation from KeepAll to KeepOne
- Clean up temp prompt files after 30s delay
- Fix badge resize_badge operator precedence

All 2302 tests pass. Rust compiles clean.

+++ src/PRDock.Tauri/src/services/diff-parser.ts
@@ -142,8 +142,8 @@ function longestCommonSubsequence(a: string, b: string): string {
   // Optimize for long strings: use two rows instead of full matrix
   if (m === 0 || n === 0) return '';
 
-  const prev = new Uint16Array(n + 1);
-  const curr = new Uint16Array(n + 1);
+  const prev = new Uint32Array(n + 1);
+  const curr = new Uint32Array(n + 1);
 
   // Build LCS length table
   const directions = new Uint8Array(m * n);
@@ -189,8 +189,8 @@ function computeSimilarity(a: string, b: string): number {
   let common = 0;
   const m = a.length;
   const n = b.length;
-  const prev = new Uint16Array(n + 1);
-  const curr = new Uint16Array(n + 1);
+  const prev = new Uint32Array(n + 1);
+  const curr = new Uint32Array(n + 1);
 
   for (let i = 1; i <= m; i++) {
     for (let j = 1; j <= n; j++) {
